【新增】配置

+++ microservice/bookstore/src/main/java/cn/hebidu/microservice/bookstore/repo/BsBookCategoryRepository.java
@@ -3,6 +3,6 @@ package cn.hebidu.microservice.bookstore.repo;
 import cn.hebidu.microservice.bookstore.entities.BsBookCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BsBookCategoryRepository extends JpaRepository<BsBookCategoryEntity, Integer> {
+public interface BsBookCategoryRepository extends JpaRepository<BsBookCategoryEntity, Long> {
 
 }
